Move code to separate folder
