+++ ydb/core/kqp/compute_actor/kqp_compute_actor_factory.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <ydb/core/protos/tx_datashard.pb.h>
 #include <ydb/library/actors/core/actor.h>
 #include <ydb/library/accessor/accessor.h>
@@ -6,7 +8,7 @@
 #include <ydb/library/yql/dq/actors/compute/dq_compute_actor.h>
 #include <ydb/core/kqp/rm_service/kqp_rm_service.h>
 
-#include <ydb/core/kqp/runtime/kqp_compute_scheduler.h>
+#include <ydb/core/kqp/runtime/scheduler/kqp_compute_scheduler.h>
 
 #include <vector>
